+++ src/modes/three_strikes_battle.cpp
@@ -660,7 +660,7 @@ void ThreeStrikesBattle::spawnSpareTireKarts()
     if (ratio < 1.5f) return;
     unsigned int spawn_sta = unsigned(ratio);
     if (spawn_sta > m_spare_tire_karts.size())
-        spawn_sta = m_spare_tire_karts.size();
+        spawn_sta = (int)m_spare_tire_karts.size();
     m_race_gui->addMessage(_P("%i spare tire kart has been spawned!",
                               "%i spare tire karts have been spawned!",
                               spawn_sta), NULL, 2.0f);
@@ -719,15 +719,15 @@ void ThreeStrikesBattle::loadCustomModels()
 
             // Compute a random kart list
             std::vector<std::string> sta_list;
-            kart_properties_manager->getRandomKartList(pos.size(), NULL,
-                &sta_list);
+            kart_properties_manager->getRandomKartList((int)pos.size(), NULL,
+                                                       &sta_list);
 
             assert(sta_list.size() == pos.size());
             // Now add them
             for (unsigned int i = 0; i < pos.size(); i++)
             {
-                AbstractKart* sta = new Kart(sta_list[i], m_karts.size(),
-                    m_karts.size() + 1, pos[i], PLAYER_DIFFICULTY_NORMAL,
+                AbstractKart* sta = new Kart(sta_list[i], (int)m_karts.size(),
+                    (int)m_karts.size() + 1, pos[i], PLAYER_DIFFICULTY_NORMAL,
                     KRT_RED);
                 sta->init(RaceManager::KartType::KT_SPARE_TIRE);
                 sta->setController(new SpareTireAI(sta));
